+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Update support for RGB Power Switch (`dj`).
 - Support showing device online status via `StatusActive`. (#172)
 - Update unit and range of `RotationSpeed` with level, need clean accessory cache to take effect. (#174)
+- Support Door and Window Controller motor reverse mode.
 
 
 ## [1.6.0] - (2022.12.3)